Version 2.1
* NEW Weekend mode - One button press prevents the smart alarm going
off and automatically re-enables it for the next night. Ideal for those
nights when you go to bed knowing you don't need to get up. Activated
by a long press on the down button either before or after the long
press on the reset button.
* Further battery optimisation - calculation for smart alarm threshold
was taking place outside of the smart alarm period. This was a waste.
* Interfacing of the smart alarm actual time is now done as part of the
assured delivery loop. This prevents the smart alarm actual time being
lost if the watch app is shut when comms is down.
* New menu icon - again. Not satisfied with the last one.
* Reset does a double buzz if the smart alarm is on and a single buzz
if it is off. As before a total failure gives an SOS buzz.
* Built with latest SDK
* Comms icon cannot be at odds with BT icon