Version 2.2
1) Ignore - a double press of the select button marks the current 10
minute period as 'ignored' (this can be undone by another double
press). Ignored periods appear as a vertical grey bar on the chart and
are included in their own (grey) segment on the pie chart. The idea is
that getting up to visit the loo, get a drink or a little nighttime
tweeting doesn't get included in the sleep pattern. There is a little
'ignore' icon on the top of the watch display.
2) The comms activity has been reduced to an absolute minimum - this
results in noticeably better battery life.
3) The are three links on the bottom of the chart page to instructions
at www.morpheuz.net, and links to the facebook page and twitter profile
apps via the url scheme. The facebook and twitter are text only on
Android.
4) Notice screen has been replaced - it isn't as artistic but I needed
more space for internationalisation of the text (English is the most
terse of the languages so room for longer text was essential). This is
really preparation for the future.
5) Animated initial screen - a little fun (I hope).
6) Integer arithmetic throughout. This has a very significant code size
change, but should also have a minor positive impact on battery life.
7) Resend data (double press of back button) - Morpheuz sends all data
to the phone ready for the settings button to be pressed. It uses the
ACK from the phone to decide that the transfer has been completed.
Sometimes the javascript on the phone doesn't run and there are gaps in
the chart. Pressing resend will re-transmit all of the data from the
watch to the phone over the next minute or so.
8) The keyboard hint page is no longer a graphic - the text is Pebble
generated - again this is for internationalisation. It doesn't look
much different.
9) The app icon has changed again. Honestly this is the last time - I'm
done with low res 1 bit plane images.