Version 3.5
1) The menu doesn't show "Ignore" until Morpheuz is recording sleep.
The top menu item is "Bed time" I.e. The first thing you might want to
select.
2) The Smart Alarm dialog has been replaced with something similar to
the quiet time setting screens from the 3.x OS. Pressing select moves
to the next field and up and down change the value. If the smart alarm
is off at the end of this, then it will be turned on.
3) The smart alarm on/off is a separate menu item. Picking this item
turns it on or off. It is always turned on after picking times in the
Smart Alarm dialog, or picking a preset.
4) The menu closes after most options, except turning the smart alarm
on or off, setting a new smart alarm time, or accessing the preset
menu. The reasoning being that it is likely that another menu item will
be picked after these ones. Use cases like set smart alarm times, then
pick bed time. Or set times and store in preset memory.
5) All watch face times now pay correct attention to 12hr or 24hr
setting.
6) The smart alarm setting menu item shows the time range picked in the
description.
7) The date on the tops of menus and the main screen includes the day
of the week. I really care this is Friday night....
9) The IFTTT Maker interface now gets a morpheuz_bedtime event too.
10) The app starts faster on Pebble and Pebble Steel.
11) The font size for setting the alarm times is now big enough for 40+
year olds to read.
12) Little notices to tell you to press "Bed time" don't appear whilst
you are on the menu trying to press the "Bed time" option. This seemed
rude.
13) The times for the smart alarm can straddle midnight.
14) A typo showing hyphens on the IFTTT maker events on the settings
page has been fixed. Should be underscores.
15) Alarm button markers appear when the alarm sounds. No more guessing
which button to press.