Version 3.5
Removed Weekend mode. It is now so easy to turn the alarm on and off,
that the weekend mode (alarm is on, but it isn’t for 12 hours) isn’t
needed. It is a source of confusion only (only emails I could find
referencing it were confused by it). It was added when alarms were set
from the phone, so provided a watch side override.